+++ platform-operator/controllers/verrazzano/component/opensearch/opensearch.go
@@ -102,7 +102,7 @@ func isOSNodeReady(ctx spi.ComponentContext, node vzapi.OpenSearchNode, prefix s
 	nodeControllerName := getNodeControllerName(node, isLegacyOS)
 
 	// If a node has the master role, it is a statefulset
-	// If the operator is managing OS, then all nodes are statefulset
+	// If the opster operator is managing OpenSearch, then all nodes are statefulset
 	if !isLegacyOS || hasRole(node.Roles, vmov1.MasterRole) {
 		return ready.StatefulSetsAreReady(ctx.Log(), ctx.Client(), []types.NamespacedName{{
 			Name:      nodeControllerName,